Performace issue
Use `qty[-1]` instead of `qty[len(qty) - 1]` to get the last element
Use `names[-1]` instead of `names[len(names) - 1]` to get the last element
Use `sl[-1]` instead of `sl[len(sl) - 1]` to get the last element

There’s no need to calculate the length of an iterable in order to fetch the last element of the iterable. You can provide a negative index -1 to it directly in order to get the last element. In this way, you don't have to iterate over the sequence using len to get the last index when your purpose is only to get the last element.